src/requirements/requirements.txt
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-extensions==3.1.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
pytest-pythonpath==0.7.3
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
